# docs/swagger.yaml
definitions:
...
  user_dtos.UserWithoutSensitiveData:
    properties:
      email:
        example: fulano@fulano.com
        type: string
      id:
        example: 1
        type: integer
      inactive:
        example: false
        type: boolean
      isAdmin:
        example: true
        type: boolean
    type: object
  utils_interfaces.ErrorResponse:
    properties:
      error:
        example: variable error message
        type: string
      statusCode:
        example: -1
        type: integer
    type: object
info:
  contact: {}
  description: Authentication API
  title: API
  version: "1.0"
paths:
...
  /users/changeEmail/{find}:
    patch:
      description: Change Email by id or email
      parameters:
      - description: 'Search parameter: e-mail or id'
        in: path
        name: find
        required: true
        type: string
      - default: Bearer <token>
        description: JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_dtos.UserWithoutSensitiveData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils_interfaces.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils_interfaces.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils_interfaces.ErrorResponse'
      security:
      - Bearer: []
      summary: Change Email (You will need send a JWT token in authorization header,
        you can get it in the login route)
      tags:
      - user
  /users/changePassword/{find}:
    patch:
      description: Change Password by id or email
      parameters:
      - description: 'Search parameter: e-mail or id'
        in: path
        name: find
        required: true
        type: string
      - default: Bearer <token>
        description: JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_dtos.UserWithoutSensitiveData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils_interfaces.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils_interfaces.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils_interfaces.ErrorResponse'
      security:
      - Bearer: []
      summary: Change Password (You will need send a JWT token in authorization header,
        you can get it in the login route)
      tags:
      - user
...
swagger: "2.0"